backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
mysqlclient>=2.1.1
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
django-filter>=23.5